🚚 [RUMF-902] move `MutationController` in `mutationObserver`
